+++ spring-project-06-api/src/main/java/jp/co/solxyz/jsn/sample/SampleApplication.java
