+++ spring-project-06-api/src/main/java/jp/co/solxyz/jsn/sample/service/BooksService.java
@@ -1,0 +1,33 @@
+package jp.co.solxyz.jsn.sample.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+import jp.co.solxyz.jsn.sample.dto.BooksDto;
+
+@Service
+public class BooksService {
+
+	/** Rest Template */
+    private final RestTemplate restTemplate;
+
+
+    @Autowired
+    public BooksService(RestTemplateBuilder builder) {
+        this.restTemplate = builder.build();
+    }
+
+    /** 通信先URL */
+    private static final String URL = "http://private-bde553-bookmanagementapi.apiary-mock.com/books";
+
+	/**
+	 * 書籍情報の取得
+	 * @return 取得した書籍情報のリスト
+	 */
+	public BooksDto[] getBooks() {
+        return restTemplate.getForObject(URL, BooksDto[].class);
+    }
+
+}
