+++ spring-project-06-api/src/main/resources/templates/booklist.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+<meta charset="UTF-8">
+<title>書籍一覧</title>
+</head>
+<body>
+<p th:text="書籍一覧"></p>
+<table th:each="book : ${books}" border="1">
+<tr><th>ID</th><th>タイトル</th><th>著者</th><th>出版社</th></tr>
+<tr><td th:text="${book.id}"></td>
+    <td th:text="${book.title}"></td>
+    <td th:text="${book.author}"></td>
+    <td th:text="${book.publisher}"></td>
+</tr>
+</table>
+</body>
+</html>
