+++ spring-project-06-api/src/test/java/jp/co/solxyz/jsn/sample/SampleApplicationTests.java
